refactor: Terminal 파일 분리
[#69]
