+++ packages/frontend/src/components/terminal/Prompt.tsx
@@ -1,0 +1,5 @@
+import { prompt as promptStyle } from "./Terminal.css";
+
+export default function Prompt() {
+  return <span className={promptStyle}>$</span>;
+}
